@@ -3,7 +3,7 @@ import { BackendHostName, FrontendHostName } from './HostName';
 const setBackendHost = (): string => {
     const hostname = window && window.location && window.location.hostname;
 
-    console.log(JSON.stringify(process.env.REACT_APP_BACKEND_SERVICE_HOST));
+    console.log(JSON.stringify(process.env));
 
     switch (hostname) {
         case FrontendHostName.PRODUCTION:
@@ -14,12 +14,13 @@ const setBackendHost = (): string => {
             return BackendHostName.TEST;
         default:
             return (
-                `${process.env.REACT_APP_BACKEND_HOST}` ||
-                'https://localhost:5001'
+                `${process.env.REACT_APP_API_URL}` || `${hostname}` || 'https://localhost:5001'
             );
     }
 };
 
 const backendHost: string = setBackendHost();
+
+console.log(backendHost);
 
 export const API_ROOT = `${backendHost}`;
